@@ -1,65 +1,67 @@
 {
-    "name": "angular",
-    "version": "0.0.0",
-    "license": "MIT",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "test": "ng test --single-run",
-        "lint": "ng lint",
-        "e2e": "ng e2e",
-        "pre-commit": "lint-staged"
-    },
-    "pre-commit": [
-        "pre-commit"
-    ],
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "^4.4.4",
-        "@angular/cdk": "^2.0.0-beta.12",
-        "@angular/common": "^4.0.0",
-        "@angular/compiler": "^4.0.0",
-        "@angular/core": "^4.0.0",
-        "@angular/forms": "^4.0.0",
-        "@angular/http": "^4.0.0",
-        "@angular/material": "^2.0.0-beta.12",
-        "@angular/platform-browser": "^4.0.0",
-        "@angular/platform-browser-dynamic": "^4.0.0",
-        "@angular/router": "^4.0.0",
-        "@types/jquery": "^3.2.16",
-        "core-js": "^2.4.1",
-        "jquery": "^3.2.1",
-        "rxjs": "^5.1.0",
-        "zone.js": "^0.8.4"
-    },
-    "devDependencies": {
-        "@angular/cli": "^1.0.1",
-        "@angular/compiler-cli": "^4.0.0",
-        "@types/jasmine": "2.5.38",
-        "@types/jquery": "^3.2.16",
-        "@types/node": "~6.0.60",
-        "codelyzer": "~2.0.0",
-        "jasmine-core": "~2.5.2",
-        "jasmine-spec-reporter": "~3.2.0",
-        "karma": "~1.4.1",
-        "karma-chrome-launcher": "~2.0.0",
-        "karma-cli": "~1.0.1",
-        "karma-coverage-istanbul-reporter": "^0.2.0",
-        "karma-jasmine": "~1.1.0",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "lint-staged": "^4.3.0",
-        "pre-commit": "^1.2.2",
-        "prettier": "^1.8.2",
-        "protractor": "~5.1.0",
-        "ts-node": "~2.0.0",
-        "tslint": "~4.5.0",
-        "typescript": "~2.2.0"
-    },
-    "lint-staged": {
-        "angular/app/**/*.ts": [
-            "prettier --single-quote --trailing-comma es5 --print-width 100 --write",
-            "git add"
-        ]
-    }
+  "name": "angular",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test --single-run",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "pre-commit": "lint-staged"
+  },
+  "pre-commit": [
+    "pre-commit"
+  ],
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^7.2.14",
+    "@angular/cdk": "^7.3.7",
+    "@angular/common": "^7.2.14",
+    "@angular/compiler": "^7.2.14",
+    "@angular/core": "^7.2.14",
+    "@angular/forms": "^7.2.14",
+    "@angular/http": "^7.2.14",
+    "@angular/material": "^7.3.7",
+    "@angular/platform-browser": "^7.2.14",
+    "@angular/platform-browser-dynamic": "^7.2.14",
+    "@angular/router": "^7.2.14",
+    "@types/jquery": "^3.2.16",
+    "core-js": "^2.4.1",
+    "jquery": "^3.2.1",
+    "rxjs": "^6.5.1",
+    "rxjs-compat": "^6.5.1",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.8.4"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "^7.3.8",
+    "@angular/compiler-cli": "^7.2.14",
+    "@types/jasmine": "2.5.38",
+    "@types/jquery": "^3.2.16",
+    "@types/node": "~6.0.60",
+    "jasmine-core": "~2.5.2",
+    "jasmine-spec-reporter": "~3.2.0",
+    "karma": "^4.1.0",
+    "karma-chrome-launcher": "~2.0.0",
+    "karma-cli": "~1.0.1",
+    "karma-coverage-istanbul-reporter": "^0.2.0",
+    "karma-jasmine": "~1.1.0",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "lint-staged": "^4.3.0",
+    "pre-commit": "^1.2.2",
+    "prettier": "^1.8.2",
+    "protractor": "^6.0.0",
+    "ts-node": "~2.0.0",
+    "tslint": "~4.5.0",
+    "typescript": "~3.2.4"
+  },
+  "lint-staged": {
+    "angular/app/**/*.ts": [
+      "prettier --single-quote --trailing-comma es5 --print-width 100 --write",
+      "git add"
+    ]
+  }
 }
